+++ lib_xua/lib_build_info.cmake
@@ -19,12 +19,16 @@ set(LIB_INCLUDES api
                  src/core/user/hostactive
                  src/hid
                  src/midi)
-set(LIB_OPTIONAL_HEADERS xua_conf.h static_hid_report.h)
+set(LIB_OPTIONAL_HEADERS    xua_conf.h
+                            static_hid_report.h
+                            user_main_globals.h
+                            user_main_declarations.h
+                            user_main_cores.h)
 set(LIB_DEPENDENT_MODULES "lib_adat(2.0.1)"
                           "lib_locks(2.3.1)"
                           "lib_logging(3.3.1)"
                           "lib_spdif(6.2.1)"
-                          "lib_sw_pll(2.3.1)"
+                          "lib_sw_pll(develop)"
                           "lib_xassert(4.3.1)"
                           "lib_mic_array(5.5.0)"
                           "lib_xud(develop)")